Add Memgraph MCP Server to README
We started working on the Memgraph MCP Server, so I added the link here to make it available to developers looking for a graph database as a part of their stack. 

Currently, the server has a tool to run a query and a schema resource. The plan is to extend it with other tools.